requirements.txt
ollama
langchain
langchain_ollama
pillow
flask
flask_cors